split `tests.rs` into multiple files
